simplifies and moves inspect_h5 to utils.py
